@trilinos/tpetra

## Motivation
- Add `isLocallyFitted` to `Details::Transfer` (and hence `Import` and `Export`). This can be cheaper than calling `Map`'s `isLocallyFitted`, since comparison of local indices is avoided.
- Skip `copyAndPermute` in `Details::residual` if domain and column map are locally fitted (OFF by default).
- Add method `computeLocalOffRankOffsets` to `CrsGraph`.
- Use Kokkos Kernels SpMV launch parameters for Tpetra's residual, they currently are a copy anyways, except for the HIP part that was recently added to Kokkos Kernels.
- Implement communication/computation overlap for Tpetra's residual (OFF by default).
- Add environment variables `TPETRA_SKIP_COPY_AND_PERMUTE` and `TPETRA_OVERLAP`

## Timing results
I ran `MueLu_Driver.exe` on 2 nodes, 8 GPUs of Lassen, UVM=OFF:
```
jsrun -M "-gpu" -p 8 -r 4 -g1 -c4 -brs ./MueLu_Driver.exe --nx=160 --ny=160 --nz=80 --matrixType=Laplace3D --resolve=19 --stacked-timer

--------------------------------------------------------------------------------
---                            Multigrid Summary Laplace3D                   ---
--------------------------------------------------------------------------------
Scalar              = double
Number of levels    = 4
Operator complexity = 1.58
Smoother complexity = 1.74
Cycle type          = V

level  rows     nnz       nnz/row  c ratio  procs
  0  2048000  14233600  6.95                  8  
  1  258240   7802664   30.21    7.93         8  
  2  6395     450867    70.50    40.38        1  
  3  90       4150      46.11    71.06        1  

Smoother (level 0) both : "Ifpack2::Relaxation": {Initialized: true, Computed: true, Type: Jacobi, sweeps: 4, damping factor: 1, Global matrix dimensions: [2048000, 2048000], Global nnz: 14233600}

Smoother (level 1) both : "Ifpack2::Relaxation": {Initialized: true, Computed: true, Type: Jacobi, sweeps: 4, damping factor: 1, Global matrix dimensions: [258240, 258240], Global nnz: 7802664}

Smoother (level 2) both : "Ifpack2::Relaxation": {Initialized: true, Computed: true, Type: Jacobi, sweeps: 4, damping factor: 1, Global matrix dimensions: [6395, 6395], Global nnz: 450867}

Smoother (level 3) pre  : KLU2 solver interface
Smoother (level 3) post : no smoother
```
I switched the smoother to Jacobi, since it uses `Tpetra::Details::residual` and switched to the `Isend` comm option.
The rows/GPU roughly corresponds to the regime EMPIRE is interested in.

I ran 7 different modes:
- using CuSparse (after setting `TPETRA_DETAILS_USE_REFERENCE_RESIDUAL`)
- using KokkosKernels SpMV (after setting `TPETRA_DETAILS_USE_REFERENCE_RESIDUAL`)
- without skipping `unpackAndCombine` (as added in #9133),
- with skipping `unpackAndCombine` (as added in #9133),
- with skipping `unpackAndCombine` and skipping `copyAndPermute`,
- with skipping `unpackAndCombine` and skipping `copyAndPermute` and overlapping of communication & computation,
- with skipping `unpackAndCombine` and skipping `copyAndPermute` and overlapping of communication & computation and persistent MPI requests #4782.

```
Belos: PseudoBlockCGSolMgr total solve time: 2.43972 - 95.7164% [20] {min=2.43859, max=2.44026, std dev=0.000710518} <2, 0, 0, 0, 0, 0, 1, 0, 1, 4>
Belos: PseudoBlockCGSolMgr total solve time: 2.68613 - 96.3204% [20] {min=2.68458, max=2.68874, std dev=0.00139822} <2, 1, 0, 1, 1, 2, 0, 0, 0, 1>
Belos: PseudoBlockCGSolMgr total solve time: 2.5866 - 96.1485% [20] {min=2.58586, max=2.58746, std dev=0.000520589} <1, 1, 0, 1, 2, 0, 1, 1, 0, 1>
Belos: PseudoBlockCGSolMgr total solve time: 2.49995 - 95.6433% [20] {min=2.49809, max=2.50577, std dev=0.00243848} <2, 3, 2, 0, 0, 0, 0, 0, 0, 1>
Belos: PseudoBlockCGSolMgr total solve time: 2.15975 - 94.9339% [20] {min=2.15824, max=2.16495, std dev=0.00229803} <4, 2, 0, 0, 1, 0, 0, 0, 0, 1>
Belos: PseudoBlockCGSolMgr total solve time: 2.17401 - 94.354% [20] {min=2.17236, max=2.17842, std dev=0.00197005} <2, 4, 0, 0, 1, 0, 0, 0, 0, 1>
Belos: PseudoBlockCGSolMgr total solve time: 2.04587 - 94.5417% [20] {min=2.04446, max=2.05102, std dev=0.00222172} <4, 2, 0, 1, 0, 0, 0, 0, 0, 1>
```
Looking at the residual in isolation:
```
MueLu: Laplace3D: Hierarchy: Solve : residual calculation (level=0): 0.0803768 - 8.82234% [360] {min=0.0801641, max=0.0804999, std dev=0.00010464} <1, 0, 0, 1, 0, 0, 2, 2, 1, 1>
MueLu: Laplace3D: Hierarchy: Solve : residual calculation (level=0): 0.109894 - 9.6762% [360] {min=0.10946, max=0.110719, std dev=0.000386159} <1, 2, 1, 2, 1, 0, 0, 0, 0, 1>
MueLu: Laplace3D: Hierarchy: Solve : residual calculation (level=0): 0.103716 - 9.60062% [360] {min=0.103493, max=0.104001, std dev=0.000160544} <1, 0, 2, 1, 1, 1, 0, 1, 0, 1>
MueLu: Laplace3D: Hierarchy: Solve : residual calculation (level=0): 0.100183 - 9.75137% [360] {min=0.0998461, max=0.100391, std dev=0.000189245} <1, 0, 1, 0, 1, 0, 1, 1, 1, 2>
MueLu: Laplace3D: Hierarchy: Solve : residual calculation (level=0): 0.0787843 - 9.1647% [360] {min=0.0785859, max=0.079288, std dev=0.000235638} <3, 2, 0, 1, 1, 0, 0, 0, 0, 1>
MueLu: Laplace3D: Hierarchy: Solve : residual calculation (level=0): 0.0783778 - 9.03264% [360] {min=0.0777182, max=0.0788482, std dev=0.000395684} <1, 0, 0, 2, 1, 0, 0, 1, 1, 2>
MueLu: Laplace3D: Hierarchy: Solve : residual calculation (level=0): 0.0757487 - 9.03905% [360] {min=0.0753943, max=0.0759687, std dev=0.000204847} <1, 1, 0, 0, 0, 0, 1, 2, 2, 1>
```
It looks like it might be worth to enable skipping of `copyAndPermute` by default.
I'm not really sure communication & computation overlap is worth it, but there might be some room for improvement.